+++ packages/components/src/components/assets/CoinLogo/coins.ts
@@ -15,7 +15,6 @@ export const COINS = {
     nmc: require('../../../images/coins/nmc.svg'),
     tada: require('../../../images/coins/tada.svg'),
     thol: require('../../../images/coins/thol.svg'),
-    tgor: require('../../../images/coins/tgor.svg'),
     tsep: require('../../../images/coins/tsep.svg'),
     txrp: require('../../../images/coins/txrp.svg'),
     vtc: require('../../../images/coins/vtc.svg'),