+++ suite-common/icons/src/icons.ts
@@ -128,7 +128,6 @@ export const cryptoIcons = {
     sol: require('../assets/cryptoIcons/sol.svg'),
     tada: require('../assets/cryptoIcons/tada.svg'),
     test: require('../assets/cryptoIcons/test.svg'),
-    tgor: require('../assets/cryptoIcons/tgor.svg'),
     thol: require('../assets/cryptoIcons/thol.svg'),
     tsep: require('../assets/cryptoIcons/tsep.svg'),
     txrp: require('../assets/cryptoIcons/txrp.svg'),