+++ suite-native/config/src/supportedNetworks.ts
@@ -32,7 +32,7 @@ export const networkSymbolsWhitelistMap = {
         'vtc',
         'zec',
     ] as NetworkSymbol[],
-    testnet: ['test', 'regtest', 'tsep', 'tgor', 'thol', 'tada', 'txrp'] as NetworkSymbol[],
+    testnet: ['test', 'regtest', 'tsep', 'thol', 'tada', 'txrp'] as NetworkSymbol[],
 } as const satisfies Record<string, NetworkSymbol[]>;
 
 // Blacklisting coins that are allowed inside `networkSymbolsWhitelistMap` so that we don't have to configs and just filter these out