+++ suite-native/storage/src/migrations/account/v3.ts
@@ -28,7 +28,6 @@ type NetworkSymbol =
     | 'test'
     | 'regtest'
     | 'tsep'
-    | 'tgor'
     | 'thol'
     | 'txrp'
     | 'ada'