+++ packages/suite/src/storage/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Storage changelog
 
+## 44
+
+-   remove goerli accounts, txs and settings
+
 ## 43
 
 -   fixes bug cannot set properties of undefined (setting 'passwords') introduced by migration 41